python-test/requirements.txt
wasmer==1.0.0-alpha3
wasmer_compiler_cranelift==1.0.0-alpha3
Pillow==7.1.1
